[kernel][vm] add zero page optimization
Mostly a lot of shuffling around of code to better set up the page fault
and faulting path to be more flexible. Now VMOs will not read-allocate
and instead return a pointer to a single physical page in the system that is
intentionally full of zeros only.

It's not a massive optimization, but it actually forces a refactor of the
fault path to be more in lined with what happens later with true copy-on-write.

Change-Id: If20b40d7cca8efcbb2c77c111707f3f979d9a2d1